refactor(task-board): read the run env from one vault secret, drop the column
Replaces `organization_settings.task_board_env` (and the migration, adapter,
port, tool schemas, contracts, hook, settings UI and i18n that column
dragged along) with an org-scoped secret named `TASK_RUN_ENV` holding a
`.env` blob. `SecretStorage.resolveByName` and Settings → Secrets already
exist, so the whole feature is now a read + a parse + a merge.

Better than the column on two counts beyond size: the value stays
encrypted in the vault instead of riding in the org-settings payload every
member reads on shell load, and rotating it is the same flow as any other
secret.

`parseDotenv` moves from apps/web into @decocms/shared so both the env
editor's paste flow and this reader use one parser (its tests move with it).
