+++ apps/api/src/storage/types.ts
@@ -18,7 +18,6 @@ import type { ChatMessage } from "../api/routes/decopilot/types";
 import type { ProviderId, ThreadStatus } from "@decocms/shared/sdk";
 import type {
   OrgFlags,
-  TaskBoardEnvEntry,
   UserModelPreferences,
 } from "@decocms/shared/organization/schema";
 import type { ThreadMetadata } from "@decocms/shared/entities";
@@ -185,8 +184,6 @@ export interface OrganizationSettingsTable {
   // Boolean toggles bag — the flag set lives in OrgFlagsSchema
   // (@decocms/shared/organization/schema); updates shallow-merge.
   flags: JsonObject<OrgFlags> | null;
-  // Env vars every task-board run gets; vault refs only — TaskBoardEnvSchema.
-  task_board_env: JsonArray<TaskBoardEnvEntry[]> | null;
   // Virtual MCP id the org lands on (`/$org`) instead of the Super Agent.
   main_agent_id: string | null;
   createdAt: ColumnType<Date, Date | string, never>;
@@ -201,7 +198,6 @@ export interface OrganizationSettings {
   simple_mode: SimpleModeConfig | null;
   default_home_agents: DefaultHomeAgentsConfig | null;
   flags: OrgFlags | null;
-  task_board_env: TaskBoardEnvEntry[] | null;
   main_agent_id: string | null;
   createdAt: Date | string;
   updatedAt: Date | string;
