+++ apps/api/src/tools/sandbox/resolve-env.ts
@@ -39,44 +39,6 @@ export async function resolveAndPushEnv({
 }: ResolveAndPushParams): Promise<void> {
   if (!entries || entries.length === 0) return;
 
-  const env = await resolveEnvEntries({ ctx, orgId, userId, entries });
-  if (Object.keys(env).length === 0) return;
-
-  const res = await runner.proxyDaemonRequest(handle, "/_sandbox/config", {
-    method: "PUT",
-    headers: new Headers({ "content-type": "application/json" }),
-    body: JSON.stringify({ env }),
-  });
-  if (!res.ok) {
-    const body = await readBoundedText(res, CONFIG_RESPONSE_MAX_BYTES).catch(
-      () => res.statusText,
-    );
-    console.warn(
-      `[SANDBOX_START] daemon rejected env patch (${res.status}): ${body}`,
-    );
-  }
-}
-
-/**
- * Resolve env declarations (literal | secret) into a flat KEY→value map.
- *
- * ⚠️ SECURITY: the result holds secret values. Never log it.
- *
- * A secret that fails to resolve (deleted by its owner, or user-scoped to a
- * different caller) is logged and skipped — the boot continues without that key
- * rather than failing the whole start over one stale reference.
- */
-export async function resolveEnvEntries({
-  ctx,
-  orgId,
-  userId,
-  entries,
-}: {
-  ctx: StudioContext;
-  orgId: string;
-  userId: string;
-  entries: RuntimeEnvEntry[];
-}): Promise<Record<string, string>> {
   const env: Record<string, string> = {};
   for (const entry of entries) {
     if (entry.kind === "literal") {
@@ -103,5 +65,20 @@ export async function resolveEnvEntries({
       throw err;
     }
   }
-  return env;
+
+  if (Object.keys(env).length === 0) return;
+
+  const res = await runner.proxyDaemonRequest(handle, "/_sandbox/config", {
+    method: "PUT",
+    headers: new Headers({ "content-type": "application/json" }),
+    body: JSON.stringify({ env }),
+  });
+  if (!res.ok) {
+    const body = await readBoundedText(res, CONFIG_RESPONSE_MAX_BYTES).catch(
+      () => res.statusText,
+    );
+    console.warn(
+      `[SANDBOX_START] daemon rejected env patch (${res.status}): ${body}`,
+    );
+  }
 }
