+++ apps/web/src/views/settings/org-general.tsx
@@ -3,7 +3,6 @@ import { ConnectBanner } from "@/components/connect/connect-banner";
 import { OrganizationForm } from "@/components/settings/organization-form";
 import { MainAgentSettings } from "@/components/settings/main-agent-settings";
 import { ReviewSettings } from "@/components/settings/review-settings";
-import { TaskEnvSettings } from "@/components/settings/task-env-settings";
 import { NavigationSettings } from "@/components/settings/navigation-settings";
 import { DomainSettings } from "@/components/settings/domain-settings";
 import { DeleteOrganizationSection } from "@/components/settings/delete-organization-section";
@@ -22,7 +21,6 @@ export function OrgGeneralPage() {
             <OrganizationForm />
             <MainAgentSettings />
             <ReviewSettings />
-            <TaskEnvSettings />
             <NavigationSettings />
             <DomainSettings />
             <DeleteOrganizationSection />
