+++ packages/shared/src/parse-dotenv.test.ts
