+++ packages/shared/src/parse-dotenv.ts
